+++ components/sidemenu.vue
@@ -48,6 +48,10 @@
       <li v-if="hasCampus">
         <nuxt-link :to="campusLink('planificator')">
           Planificateur
+          <span
+            v-if="ridesToValidate.length > 0"
+            class="red-dot"
+          >{{ ridesToValidate.length }}</span>
         </nuxt-link>
       </li>
       <li v-if="hasCampus">
@@ -188,7 +192,10 @@ export default {
     reconnectingHero,
   },
   computed: {
-    ...mapGetters({ campus: 'context/campus' }),
+    ...mapGetters({
+      campus: 'context/campus',
+      ridesToValidate: 'realtime/ridesToValidate',
+    }),
   },
   watch: {
     campus(c) {
@@ -209,11 +216,13 @@ export default {
   methods: {
     ...mapActions({ fetchCampus: 'context/fetchCampus', setRides: 'realtime/setRides' }),
     async setCampus(campus) {
-      this.fetchCampus(campus ? campus.id : null);
+      await this.fetchCampus(campus ? campus.id : null);
       if (campus) {
         const start = DateTime.local().startOf('days').toJSDate();
-        const end = DateTime.local().endOf('days').toJSDate();
-        await this.setRides({ campus: campus.id, start, end });
+        const end = DateTime.local()
+          .plus({ seconds: this.campus.defaultReservationScope })
+          .toJSDate();
+        await this.setRides({ campus: this.campus.id, start, end });
       }
     },
     logout() {
@@ -238,6 +247,26 @@ export default {
 
   .logo {
     padding: 20px;
+  }
+
+  .red-dot {
+    position: absolute;
+    top: 16px;
+    right: 20px;
+    height: 20px;
+    width: 20px;
+    border-radius: 100%;
+    background-color: $red;
+    color: $white;
+    font-size: .7em;
+    font-weight: 700;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  a {
+    position: relative;
   }
 
   .menu-list a {