+++ store/context.js
@@ -29,7 +29,7 @@ export const actions = {
       return;
     }
     const { data } = await this.$api.campuses.getCampus(campus,
-      'id,name,location,categories(id,label),workedDays,workedHours,defaultRideDuration');
+      'id,name,location,categories(id,label),workedDays,workedHours,defaultRideDuration,defaultReservationScope');
     commit('setCampus', data);
   },
   hideMap({ commit }, hide) {